requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
transformers>=4.40.0
sentence-transformers>=2.2.0
pdfplumber>=0.10.0
PyMuPDF>=1.24.0
textstat>=0.7.0
flask>=3.0.0
flask-cors>=4.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
python-magic-bin>=0.4.14; platform_system == "Windows"
python-magic>=0.4.27; platform_system != "Windows"
numpy>=1.22.0
pandas>=2.0.0
torch>=2.0.0
